+++ app/page.tsx
@@ -4,133 +4,15 @@ import pointer from "@/public/icons/pointer.png";
 import medical from "@/public/images/hero-img.png";
 import chatBot from "@/public/images/chatBot.png";
 import HealthCard from "./component/HealthCard";
-import cancerPatient from "@/public/images/cancer-patient.png";
-import elderly from "@/public/images/elderly.png";
-import workingProfessionals from "@/public/images/workingProfessionals.png";
-import mentalPatients from "@/public/images/mentalPatient.png";
-import youngAdult from "@/public/images/youngAdults.png";
-import parents from "@/public/images/parent.png";
 import consultant from "@/public/images/consultant.png";
 import check from "@/public/icons/check.png";
-import key from "@/public/icons/key.png";
-import stethoscope from "@/public/icons/stethoscope.png";
-import record from "@/public/icons/record.png";
-import calender from "@/public/icons/calender.png";
 import Step from "./component/Step";
-
-interface HealthCardProps {
-  photo: StaticImageData;
-  title: string;
-  alt: string;
-  description: string;
-  paint: string;
-}
-
-interface StepProps {
-  icon: StaticImageData;
-  step: string;
-  title: string;
-  description: string;
-  mt: string;
-  nextArrow: boolean;
-  transform: boolean;
-}
+import { accordionData } from "./utils/Accordion";
+import Accordion from "./component/Accordion";
+import { stepCardData } from "./utils/StepCard";
+import { headerCardData } from "./utils/HeaderCard";
 
 export default function Home() {
-  const headerCardData: HealthCardProps[] = [
-    {
-      photo: cancerPatient,
-      title: "Chronic Disease Patients",
-      alt: "Chronic Disease Patients",
-      description:
-        "Manage your health with regular check-ups and support to ensure you stay on track with your treatment, because your well-being is our top priority.",
-      paint: "bg-success-10",
-    },
-    {
-      photo: elderly,
-      title: "Elderly",
-      alt: "Elderly",
-      description:
-        "Receive compassionate care right at home by using our user-friendly telehealth services, allowing you to rest easy knowing expert help is just a call away.",
-      paint: "bg-primary-10",
-    },
-    {
-      photo: workingProfessionals,
-      title: "Working Professionals",
-      alt: "Working Professionals",
-      description:
-        "Receive compassionate care right at home by using our user-friendly telehealth services, allowing you to rest easy knowing expert help is just a call away.",
-      paint: "bg-warning-10",
-    },
-    {
-      photo: mentalPatients,
-      title: "Mental Health Patients",
-      alt: "Mental Health Patients",
-      description:
-        "Access confidential counseling and support at your convenience, providing you with the tools to improve your mental well-being and first step towards a brighter tomorrow.",
-      paint: "bg-error-10",
-    },
-    {
-      photo: youngAdult,
-      title: "Young Adult",
-      alt: "Young Adult",
-      description:
-        "You can access high-quality healthcare from the comfort of your home and connect with trusted doctors whenever you need them.",
-      paint: "bg-error-20",
-    },
-    {
-      photo: parents,
-      title: "Parents",
-      alt: "Parents",
-      description:
-        "Get instant access to pediatric care to ensure your child(ren) receive the best possible attention, as we’re here to support your family's health every step of the way.",
-      paint: "bg-success-10",
-    },
-  ];
-
-  const stepCardData: StepProps[] = [
-    {
-      nextArrow: true,
-      transform: false,
-      mt: "mt-0",
-      step: "01",
-      icon: key,
-      title: "Register or Log in",
-      description:
-        "Sign up for a new account or log in to access personalized healthcare services. It's quick and easy.",
-    },
-    {
-      nextArrow: true,
-      transform: true,
-      mt: "sm:mt-[6rem]",
-      step: "02",
-      icon: stethoscope,
-      title: "Choose a Doctor",
-      description:
-        "Select the service that suits your needs, from virtual consultations to specialist appointments. We offer a range of options for you.",
-    },
-    {
-      nextArrow: true,
-      transform: false,
-      mt: "sm:mt-[12rem]",
-      step: "03",
-      icon: calender,
-      title: "Book an Appointment",
-      description:
-        "Connect instantly with a licensed doctor through secure video calls. Get expert advice and treatment from the comfort of your home.",
-    },
-    {
-      nextArrow: false,
-      transform: false,
-      mt: "sm:mt-[18rem] ",
-      step: "04",
-      icon: record,
-      title: "Manage Records",
-      description:
-        "Easily access your medical records anytime and schedule follow-ups as needed. Stay organized and in control of your health journey.",
-    },
-  ];
-
   return (
     <main className="w-full bg-primary-white flex flex-col items-center">
       {/* hero section  */}
@@ -189,14 +71,14 @@ export default function Home() {
         </div>
         {/* HealthCard Component  */}
         <div className="grid md:grid-cols-3 sm:grid-cols-2 gap-5 md:gap-10 lg:gap-20">
-          {headerCardData.map((item) => (
+          {headerCardData.map(({ alt, photo, title, description, paint }) => (
             <HealthCard
-              key={item.alt}
-              photo={item.photo}
-              title={item.title}
-              alt={item.alt}
-              description={item.description}
-              paint={item.paint}
+              key={alt}
+              photo={photo}
+              title={title}
+              alt={alt}
+              description={description}
+              paint={paint}
             />
           ))}
         </div>
@@ -253,7 +135,7 @@ export default function Home() {
       {/* How it Works */}
       <div className=" relative w-[90%] flex flex-col mt  justify-center  pb-16">
         <div className="flex justify-center sm:mb-[-6rem] mb-5 text-center ">
-          <div className="w-1/2">
+          <div className="w-[80%h] sm:w-1/2 ">
             <h3 className=" text-lg md:text-xl lg:text-2xl  font-bold ">
               How It Works
             </h3>
@@ -264,18 +146,46 @@ export default function Home() {
           </div>
         </div>
         <div className="flex flex-col sm:flex-row gap-5 md:gap-10 lg:gap-24  items-center text-center sm:text-left justify-between">
-          {stepCardData.map((item) => (
-            <Step
-              transform={item.transform}
-              key={item.step}
-              nextArrow={item.nextArrow}
-              mt={item.mt}
-              step={item.step}
-              icon={item.icon}
-              title={item.title}
-              description={item.description}
-            />
-          ))}
+          {stepCardData.map(
+            ({ transform, step, nextArrow, mt, icon, title, description }) => (
+              <Step
+                transform={transform}
+                key={step}
+                nextArrow={nextArrow}
+                mt={mt}
+                step={step}
+                icon={icon}
+                title={title}
+                description={description}
+              />
+            )
+          )}
+        </div>
+      </div>
+      {/* FAQs  */}
+      <div className="w-full pt-5 pb-16 mb-10">
+        <div className="w-full flex flex-col sm:flex-row items-center justify-between sm:mb-[-6rem] mb-5 gap-10 ">
+          <div className="flex flex-col justify-center items-center  bg-primary-base text-primary-white w-full sm:w-[45%] md:w-[40%] lg:w-[25%] h-64 sm:h-96 p-2 mb-10 sm:mb-0">
+            <div className="w-[90%] sm:w-[75%]  md:w-[65%]">
+              <h3 className="text-xl sm:text-2xl  font-bold text-center sm:text-left">
+                FAQs
+              </h3>
+              <p className="text-xs font-light py-3 ">
+                Here, we’ve compiled answers to some of the most common
+                inquiries to help you navigate your online consultation
+                experience with ease and confidence.
+              </p>
+              <button className="flex  justify-center items-center text-xs  cursor-pointer hover:scale-105 duration-500  bg-primary-white text-black  border-gray-20 py-2 px-4 rounded-xl mx-auto sm:mx-0">
+                Learn more
+                <Image src={arrow} alt="arrow" width={12} height={12} />
+              </button>
+            </div>
+          </div>
+          <div className="w-full sm:w-[70%] md:w-[60%] lg:w-[45%] flex flex-col gap-5 justify-center items-center sm:items-start">
+            {accordionData.map(({ title, content }) => (
+              <Accordion title={title} content={content} />
+            ))}
+          </div>
         </div>
       </div>
     </main>
