+++ app/utils/Accordion.ts
@@ -1,0 +1,23 @@
+interface AccordionProp {
+  title: string;
+  content: string;
+}
+
+export const accordionData: AccordionProp[] = [
+  {
+    title: "How do I schedule an appointment?",
+    content: `We offer a wide range of consultations, including primary care, mental health, dermatology, pediatrics, and more. You can browse specialties on E-Medatt to find the right fit for your needs.`,
+  },
+  {
+    title: "What types of consultations are available?",
+    content: `We offer a wide range of consultations, including primary care, mental health, dermatology, pediatrics, and more. You can browse specialties on E-Medatt to find the right fit for your needs.`,
+  },
+  {
+    title: "Is the consultation secure and private?",
+    content: `We offer a wide range of consultations, including primary care, mental health, dermatology, pediatrics, and more. You can browse specialties on E-Medatt to find the right fit for your needs.`,
+  },
+  {
+    title: "Can I change or cancel my appointment?",
+    content: `We offer a wide range of consultations, including primary care, mental health, dermatology, pediatrics, and more. You can browse specialties on E-Medatt to find the right fit for your needs.`,
+  },
+];
